+++ core/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Struct introspection (`introspection::Struct`) can now specify a special field as a fallback.
 - Service introspection (`introspection::Service`) can now specify a special function as a fallback.
 - Service introspection (`introspection::Service`) can now specify a special event as a fallback.
+- Add `SerializedValueSlice::deserialize_as_value()` to deserialize into a generic `Value`.
 
 ### Changed